configcore: fix build of snapd with -tags nomanagers
